+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### UI Improvements
 
+1.  [#3096](https://github.com/influxdata/chronograf/pull/3096): Standardize delete confirmation interactions
+1.  [#3096](https://github.com/influxdata/chronograf/pull/3096): Standardize save & cancel interactions
+
 ### Bug Fixes
 
 1.  [#2950](https://github.com/influxdata/chronograf/pull/2094): Always save template variables on first edit
@@ -58,6 +61,7 @@
 ## v1.4.2.1 [2018-02-28]
 
 ### Features
+
 1.  [#2837](https://github.com/influxdata/chronograf/pull/2837): Prevent execution of queries in cells that are not in view on the dashboard page
 1.  [#2829](https://github.com/influxdata/chronograf/pull/2829): Add an optional persistent legend which can toggle series visibility to dashboard cells
 1.  [#2846](https://github.com/influxdata/chronograf/pull/2846): Allow user to annotate graphs via UI or API